implements CR suggestions by Joao
